Task 1 Refactoring : Done
