+++ src/front-end/components/workshop.ejs
